Company crude has admin validtaions

+++ src/main/java/com/ori/api/CompanyApi.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,7 +17,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ori.controller.CompanyController;
+import com.ori.data.UserLoginData;
 import com.ori.entities.Company;
+import com.ori.enums.ErrorTypes;
+import com.ori.enums.UserType;
 import com.ori.exceptions.ApplicationException;
 
 @RestController
@@ -28,7 +33,14 @@ public class CompanyApi {
 	
 //  URL : http://localhost:8080/companies
 		@PostMapping
-		public void addCompany(@RequestBody Company company) throws ApplicationException {
+		public void addCompany(@RequestBody Company company, HttpServletRequest request) throws ApplicationException {
+			
+
+			UserLoginData userLoginData = (UserLoginData) request.getAttribute("userLoginData");
+			
+			if(userLoginData.getUserType() != UserType.ADMIN) {
+				throw new ApplicationException(ErrorTypes.UNAUTHROIZED, "UNAUTHORIZED");
+			}
 			
 			this.companyController.createCompany(company);
 			System.out.println(company);
@@ -36,8 +48,15 @@ public class CompanyApi {
 		
 		//  URL : http://localhost:8080/companies
 		@PutMapping
-		public void updateCompany(@RequestBody Company company) throws ApplicationException {
+		public void updateCompany(@RequestBody Company company, HttpServletRequest request) throws ApplicationException {
 			System.out.println(company);
+			
+
+			UserLoginData userLoginData = (UserLoginData) request.getAttribute("userLoginData");
+			
+			if(userLoginData.getUserType() != UserType.ADMIN) {
+				throw new ApplicationException(ErrorTypes.UNAUTHROIZED, "UNAUTHORIZED");
+			}
 			this.companyController.updateCompany(company);
 			System.out.println(company);
 		}
@@ -53,16 +72,28 @@ public class CompanyApi {
 		
 		// http://localhost:8080/companies/12
 		@DeleteMapping("{id}")
-		public void deleteCompany(@PathVariable("id") long id) throws ApplicationException {
+		public void deleteCompany(@PathVariable("id") long id, HttpServletRequest request) throws ApplicationException {
+			
 
+			UserLoginData userLoginData = (UserLoginData) request.getAttribute("userLoginData");
+			
+			if(userLoginData.getUserType() != UserType.ADMIN) {
+				throw new ApplicationException(ErrorTypes.UNAUTHROIZED, "UNAUTHORIZED");
+			}
+			
 			this.companyController.deleteCompany(id);
 		}
 		
 		
 		//  URL : http://localhost:8080/companies
 		@GetMapping
-		public List<Company> getAllCompanies() throws ApplicationException {
+		public List<Company> getAllCompanies(HttpServletRequest request) throws ApplicationException {
+
+			UserLoginData userLoginData = (UserLoginData) request.getAttribute("userLoginData");
 			
+			if(userLoginData.getUserType() != UserType.ADMIN) {
+				throw new ApplicationException(ErrorTypes.UNAUTHROIZED, "UNAUTHORIZED");
+			}
 			return this.companyController.getAllCompanies();
 		}
